+++ guillotina/cookiecutter/application/{{cookiecutter.package_name}}/{{cookiecutter.package_name}}/tests/conftest.py
@@ -1,4 +1,5 @@
 pytest_plugins = [
+    'aiohttp.pytest_plugin',
     'guillotina.tests.fixtures',
     '{{cookiecutter.package_name}}.tests.fixtures'
 ]